Folder restructure replacing lib with utils for utility functions
